warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@15051 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/themes.cpp
@@ -181,8 +181,8 @@ AeroEffect  CSkin::m_aeroEffects[AERO_EFFECT_LAST] = {
 		1,
 		0,
 		0xc0c0c0,
-		-1,
-		-1,
+		DWORD(-1),
+		DWORD(-1),
 		AeroEffectCallback_Solid
 	}
 };
